+++ R/zzz.R
@@ -1,7 +1,8 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.heemod <- list(
-    heemod.verbose = FALSE
+    heemod.verbose = FALSE,
+    heemod.memotime = 3600
   )
   toset <- !(names(op.heemod) %in% names(op))
   if(any(toset)) options(op.heemod[toset])